PP-30 refactor: change chat
