+++ pointing-poker/src/components/Chat/Chat.module.scss
@@ -3,43 +3,26 @@
 $greyColor: #c4d2cf;
 $whiteColor: #ffffff;
 
-@mixin scrollbars($size, $foreground-color, $background-color: mix($foreground-color, white, 50%)) {
-  &::-webkit-scrollbar {
-    width: $size;
-    height: $size;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    background: $foreground-color;
-    border-radius: 10px;
-  }
-
-  &::-webkit-scrollbar-track {
-    background: $background-color;
-    border-radius: 10px;
-  }
-}
 .chat {
-  position: absolute;
+  position: fixed;
   top: 100px;
   right: 30px;
   width: 350px;
   height: 500px;
-  background-color: $del-rio;
   padding: 1rem;
   @include flex($direction: column);
+  background-color: $del-rio;
   border-radius: 60px 45px 30px 15px/15px 30px 45px 60px;
-  box-shadow: 0 0 11px $sandy-brown-light;
-  .user-name {
-    text-align: start;
-    width: 100%;
-  }
+  box-shadow: 0 0 11px $greyColor;
+  z-index: 100;
+
   .chatMessage {
     width: 100%;
     height: 90%;
     overflow-y: auto;
-    @include scrollbars(5px, $dark-brown, $white);
     @include flex($direction: column, $align: flex-start, $justify: flex-start);
+    @include scrollbars(5px, $dark, $white);
+    font-family: sans-serif;
     word-break: break-all;
     line-height: 1;
 
@@ -49,7 +32,7 @@ $whiteColor: #ffffff;
       @include flex();
       padding-left: 0.5rem;
 
-      p {
+      .text {
         padding: 1rem;
         width: 220px;
         @include flex();
@@ -59,11 +42,11 @@ $whiteColor: #ffffff;
         color: $black;
       }
 
-      span {
+      .sender {
         padding: 0.5rem;
         width: 90px;
-        @include font($size: 1.3rem, $weight: 300);
-        color: $dark-brown;
+        @include font($size: 1rem, $weight: 300);
+        color: $black;
       }
     }
   }
@@ -72,24 +55,36 @@ $whiteColor: #ffffff;
     width: 100%;
     height: 50px;
     display: flex;
+
     input {
       width: 100%;
       text-decoration: none;
       padding-left: 1rem;
+      font-size: 16px;
+
       &:focus {
         outline: none;
       }
     }
+
     button {
       width: 20%;
       background-image: url(../../assets/icons/icon__send.svg);
       background-repeat: no-repeat;
       background-position: center;
       border-radius: 0 5px 5px 0;
+
       &:hover {
         cursor: pointer;
         transform: scale(1.1);
       }
     }
+  }
+}
+
+@media all and (max-width: 450px) {
+  .chat {
+    right: 10px;
+    width: 300px;
   }
 }
