+++ pointing-poker/src/components/Chat/ChatButton/index.ts
@@ -1,0 +1,1 @@
+export { ChatButton } from './ChatButton';
